+++ crates/swc_ecma_magicast/src/module.rs
@@ -1,8 +1,8 @@
-use std::ops::{Index, IndexMut};
+use std::ops::{Deref, Index, IndexMut};
 
 use swc_ecma_ast::Module;
 
-use crate::{data::Data, option::OptionalNode};
+use crate::{data::Data, OptionalNode};
 
 pub struct ModuleNode(Data<Module>);
 
@@ -14,6 +14,20 @@ impl ModuleNode {
 
 pub struct ModuleImports {}
 
+impl ModuleImports {
+    pub fn named(&self, name: &str) -> ImportNamed {}
+}
+
+pub struct ImportNamed {}
+
+impl Deref for ImportNamed {
+    type Target = BindingRef;
+}
+
+pub struct ImportFrom {}
+
+impl Ensurable for ImportFrom {}
+
 pub struct ModuleExports {
     pub default: OptionalNode<ExportItemNode>,
 }
@@ -23,3 +37,5 @@ impl Index<String> for ModuleExports {
 }
 
 impl IndexMut<String> for ModuleExports {}
+
+pub struct ExportItemNode {}
